ACP-L2-IMP-WRK-PANIC: Split executor-panic functional scenario into a focused test file
Move the executor_panic_routes_to_failed_terminal scenario (and its
panicInferenceProvider fake / assertDispatchResponseErrorContains
helper) out of simple_run_test.go and
TestPetriWorkerErrorReturnsFailedTerminalOutcome into a new top-level
TestPetriExecutorPanicRoutesToFailedTerminal test, so PR #1741/#1743's
panic-typing coverage stops deepening the file's/function's
pre-existing backend-size debt (simple_run_test.go returns to its
pre-#1741 1412 lines/321-line function baseline).
